feat: update resonator skill components
Since users may not always use the auto-fill feature in the echo tab, we
have included `base_attr` and `main_skill_bonus` in both the resonator
and echo tab pages.

When the resonator name is modified, update the skill as well.

Update skill damage multiplier when the skill level or skill ID is
changed.